@@ -18,7 +18,7 @@ package com.criticalblue.android.astropiks;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import java.util.UUID;
 
