@@ -16,7 +16,7 @@
 
 package com.criticalblue.android.astropiks;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 public class PhotoGalleryActivity extends SingleFragmentActivity {
 
